Renamed dialects to dialect
