+++ frictionless/dialect.py
@@ -8,7 +8,7 @@ class Dialect(Metadata):
 
     API      | Usage
     -------- | --------
-    Public   | `from frictionless import dialects`
+    Public   | `from frictionless import Dialect`
 
     Parameters:
         descriptor? (str|dict): descriptor
@@ -84,95 +84,6 @@ class Dialect(Metadata):
         "type": "object",
         "additionalProperties": False,
         "properties": {
-            "header": {"type": "boolean"},
-            "headerRows": {"type": "array", "items": {"type": "number"}},
-            "headerJoin": {"type": "string"},
-            "headerCase": {"type": "boolean"},
-        },
-    }
-
-
-class JsonDialect(Dialect):
-    """Json dialect representation
-
-    API      | Usage
-    -------- | --------
-    Public   | `from frictionless import dialects`
-
-    Parameters:
-        descriptor? (str|dict): descriptor
-        keys? (str[]): a list of strings to use as data keys
-        keyed? (bool): whether data rows are keyed
-        property? (str): a path within JSON to the data
-
-    Raises:
-        FrictionlessException: raise any error that occurs during the process
-
-    """
-
-    def __init__(
-        self,
-        descriptor=None,
-        *,
-        keys=None,
-        keyed=None,
-        property=None,
-        header=None,
-        header_rows=None,
-        header_join=None,
-        header_case=None,
-    ):
-        self.setinitial("keys", keys)
-        self.setinitial("keyed", keyed)
-        self.setinitial("property", property)
-        super().__init__(
-            descriptor=descriptor,
-            header=header,
-            header_rows=header_rows,
-            header_join=header_join,
-            header_case=header_case,
-        )
-
-    @Metadata.property
-    def keys(self):
-        """
-        Returns:
-            str[]?: keys
-        """
-        return self.get("keys")
-
-    @Metadata.property
-    def keyed(self):
-        """
-        Returns:
-            bool: keyed
-        """
-        return self.get("keyed", False)
-
-    @Metadata.property
-    def property(self):
-        """
-        Returns:
-            str?: property
-        """
-        return self.get("property")
-
-    # Expand
-
-    def expand(self):
-        """Expand metadata"""
-        super().expand()
-        self.setdefault("keyed", self.keyed)
-
-    # Metadata
-
-    metadata_profile = {  # type: ignore
-        "type": "object",
-        "additionalProperties": False,
-        "properties": {
-            "keys": {"type": "array"},
-            "keyed": {"type": "boolean"},
-            "property": {"type": "string"},
             "header": {"type": "boolean"},
             "headerRows": {"type": "array", "items": {"type": "number"}},
             "headerJoin": {"type": "string"},
