+++ frictionless/plugins/csv.py
@@ -4,7 +4,7 @@ import stringcase
 import unicodecsv
 from itertools import chain
 from ..metadata import Metadata
-from ..dialects import Dialect
+from ..dialect import Dialect
 from ..plugin import Plugin
 from ..parser import Parser
 from .. import helpers
