+++ frictionless/plugins/inline.py
@@ -1,7 +1,7 @@
 from itertools import chain
 from ..plugin import Plugin
 from ..metadata import Metadata
-from ..dialects import Dialect
+from ..dialect import Dialect
 from ..parser import Parser
 from .. import exceptions
 from .. import errors
