+++ frictionless/plugins/json.py
@@ -5,7 +5,7 @@ import simplejson
 from ..plugins.inline import InlineDialect
 from ..metadata import Metadata
 from ..resource import Resource
-from ..dialects import Dialect
+from ..dialect import Dialect
 from ..plugin import Plugin
 from ..parser import Parser
 from ..system import system
