+++ frictionless/plugins/ods.py
@@ -1,7 +1,7 @@
 import io
 from datetime import datetime
 from ..metadata import Metadata
-from ..dialects import Dialect
+from ..dialect import Dialect
 from ..plugin import Plugin
 from ..parser import Parser
 from .. import exceptions
