+++ frictionless/program/extract.py
@@ -7,9 +7,9 @@ from typing import List
 from typer import Option as Opt
 from typer import Argument as Arg
 from ..extract import extract
+from ..dialect import Dialect
 from ..query import Query
 from .main import program
-from .. import dialects
 from .. import helpers
 
 
@@ -74,7 +74,7 @@ def program_extract(
 
     # Prepare dialect
     dialect = (
-        dialects.Dialect(
+        Dialect(
             header_rows=header_rows,
             header_join=header_join,
         )
